7-days-MVP/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.1.0
rioxarray>=0.13.0
fastapi>=0.100.0
uvicorn>=0.22.0
streamlit>=1.24.0
folium>=0.14.0
jinja2>=3.1.2
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
black>=23.7.0
ruff>=0.0.284 
cdsapi>=0.7.6 